refactor: rename core to infrastructure
