+++ alembic/env.py
@@ -3,7 +3,7 @@ from alembic import context
 from sqlalchemy import engine_from_config, pool
 
 # ---- our project imports ----
-from app.core.database import engine, Base  # uses the same engine as your app
+from app.infrastructure.database import engine, Base  # uses the same engine as your app
 from app import models  # noqa: F401
 
 # Alembic Config object
