+++ app/models/task.py
@@ -7,7 +7,7 @@ from sqlalchemy import (
 )
 from sqlalchemy.orm import relationship
 
-from app.core.database import Base
+from app.infrastructure.database import Base
 
 
 class Priority(str, enum.Enum):
