+++ app/routers/notifications.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends, Query
 from sqlalchemy import select, update
 from sqlalchemy.orm import Session
 
-from app.core.database import get_db
+from app.infrastructure.database import get_db
 from app.models.notification import Notification
 from app.models.user import User
 from app.routers.users import get_current_user
